# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI_Chatbot API Skeleton
  version: 0.1.0-skeleton
  description: Auto-generated skeleton from workbook; schemas are intentionally coarse.
servers:
- url: http://localhost:8080
paths:
...
  /v1/admin/change-notices/{notice_id}/publish:
    post:
      operationId: opsChangeNoticePublish
      summary: 변경 공지 발행
      description: 변경 공지 발행 기능 제공
      tags:
      - 운영 관측 통합
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: OPS-CHANGE-NOTICE-PUBLISH
      x-source-row: 59
...
  /v1/admin/deploy-approvals:
    post:
      operationId: opsDeployApprovalCreate
      summary: 배포 승인 요청
      description: 배포 승인 요청 기능 제공
      tags:
      - 운영 관측 통합
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: OPS-DEPLOY-APPROVAL-CREATE
      x-source-row: 57
  /v1/admin/deploy-approvals/{approval_id}/actions/{action}:
    post:
      operationId: opsDeployApprovalAct
      summary: 배포 승인 처리
      description: 배포 승인 처리 기능 제공
      tags:
      - 운영 관측 통합
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
      - name: action
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: OPS-DEPLOY-APPROVAL-ACT
      x-source-row: 58
  /v1/admin/errors/catalog/{error_code}:
    put:
      operationId: opsErrorCatalogUpsert
      summary: 에러코드 카탈로그 관리
      description: 에러코드 카탈로그 관리 기능 제공
      tags:
      - 운영 관측 통합
      parameters:
      - name: error_code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: OPS-ERROR-CATALOG-UPSERT
      x-source-row: 56
...
  /v1/admin/models:
    get:
      operationId: admModelList
      summary: 모델 목록 조회
      description: 모델 목록 조회 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-MODEL-LIST
      x-source-row: 38
    post:
      operationId: admModelCreate
      summary: 모델 등록
      description: 모델 등록 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-MODEL-CREATE
      x-source-row: 39
  /v1/admin/models/{model_id}/activate:
    post:
      operationId: admModelActivate
      summary: 활성 모델 전환
      description: 활성 모델 전환 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-MODEL-ACTIVATE
      x-source-row: 40
  /v1/admin/models/{model_id}/rollback:
    post:
      operationId: admModelRollback
      summary: 활성 모델 롤백
      description: 활성 모델 롤백 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-MODEL-ROLLBACK
      x-source-row: 41
  /v1/admin/policies/{policy_id}:
    put:
      operationId: admPolicyUpdate
      summary: 정책/프롬프트 업데이트
      description: 정책/프롬프트 업데이트 기능 제공
      tags:
      - 템플릿 정책 관리
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-POLICY-UPDATE
      x-source-row: 37
  /v1/admin/provider-keys/{provider}:
    put:
      operationId: admProviderKeyUpsert
      summary: Provider 키 등록/수정
      description: Provider 키 등록/수정 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-PROVIDER-KEY-UPSERT
      x-source-row: 44
  /v1/admin/provider-keys/{provider}/rotate:
    post:
      operationId: admProviderKeyRotate
      summary: Provider 키 회전
      description: Provider 키 회전 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-PROVIDER-KEY-ROTATE
      x-source-row: 45
  /v1/admin/providers/{provider_id}/secret-ref:
    post:
      operationId: admProviderSecretrefBind
      summary: Provider secret_ref bind
      description: Bind/rotate provider secret reference (never plaintext key).
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-PROVIDER-SECRETREF-BIND
      x-source-row: 80
...
  /v1/admin/routing-rules/test:
    post:
      operationId: admRoutingRuleTest
      summary: 라우팅 룰 테스트
      description: 라우팅 룰 테스트 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-ROUTING-RULE-TEST
      x-source-row: 43
  /v1/admin/routing-rules/{rule_id}:
    put:
      operationId: admRoutingRuleUpsert
      summary: 라우팅 룰 등록/수정
      description: 라우팅 룰 등록/수정 기능 제공
      tags:
      - LLM 모델 시크릿 관리
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: ADM-ROUTING-RULE-UPSERT
      x-source-row: 42
  /v1/admin/templates:
    get:
      operationId: tmpTemplateList
      summary: 템플릿 목록 조회
      description: 템플릿 목록 조회 기능 제공
      tags:
      - 템플릿 정책 관리
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: TMP-TEMPLATE-LIST
      x-source-row: 31
    post:
      operationId: tmpTemplateCreate
      summary: 템플릿 생성
      description: 템플릿 생성 기능 제공
      tags:
      - 템플릿 정책 관리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: TMP-TEMPLATE-CREATE
      x-source-row: 32
  /v1/admin/templates/{template_id}/approve:
    post:
      operationId: tmpTemplateApprove
      summary: 템플릿 승인
      description: 템플릿 승인 기능 제공
      tags:
      - 템플릿 정책 관리
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: TMP-TEMPLATE-APPROVE
      x-source-row: 33
  /v1/admin/templates/{template_id}/deploy:
    post:
      operationId: tmpTemplateDeploy
      summary: 템플릿 배포
      description: 템플릿 배포 기능 제공
      tags:
      - 템플릿 정책 관리
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: TMP-TEMPLATE-DEPLOY
      x-source-row: 34
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT